REV-006: extract shared review patch cache keys (s17) (#17)
Move fnv1a32 and buildPatchCacheKey to @t3tools/shared/review so web
diffRendering and mobile reviewModel share identical Pierre parse cache keys.
