+++ apps/mobile/src/features/review/reviewModel.ts
@@ -1,6 +1,7 @@
 import { parsePatchFiles } from "@pierre/diffs/utils/parsePatchFiles";
 import type { ChangeTypes, FileDiffMetadata } from "@pierre/diffs/types";
 import type { OrchestrationCheckpointSummary, ReviewDiffPreviewSource } from "@t3tools/contracts";
+import { buildPatchCacheKey } from "@t3tools/shared/review";
 import * as Arr from "effect/Array";
 import { pipe } from "effect/Function";
 import * as Order from "effect/Order";
@@ -204,10 +205,6 @@ function runDiffParserSilently<T>(callback: () => T): T {
   }
 }
 
-const FNV_OFFSET_BASIS_32 = 0x811c9dc5;
-const FNV_PRIME_32 = 0x01000193;
-const SECONDARY_HASH_SEED = 0x9e3779b9;
-const SECONDARY_HASH_MULTIPLIER = 0x85ebca6b;
 const LARGE_DIFF_LINE_THRESHOLD = 400;
 const LARGE_DIFF_CHARACTER_THRESHOLD = 24_000;
 const NON_TEXT_FILE_EXTENSIONS = new Set([
@@ -252,26 +249,6 @@ const NON_TEXT_FILE_EXTENSIONS = new Set([
   "so",
   "dylib",
 ]);
-
-function fnv1a32(input: string, seed: number, multiplier: number): number {
-  let hash = seed >>> 0;
-  for (let index = 0; index < input.length; index += 1) {
-    hash ^= input.charCodeAt(index);
-    hash = Math.imul(hash, multiplier) >>> 0;
-  }
-  return hash >>> 0;
-}
-
-function buildPatchCacheKey(patch: string, scope: string): string {
-  const normalizedPatch = patch.trim();
-  const primary = fnv1a32(normalizedPatch, FNV_OFFSET_BASIS_32, FNV_PRIME_32).toString(36);
-  const secondary = fnv1a32(
-    normalizedPatch,
-    SECONDARY_HASH_SEED,
-    SECONDARY_HASH_MULTIPLIER,
-  ).toString(36);
-  return `${scope}:${normalizedPatch.length}:${primary}:${secondary}`;
-}
 
 function getFileExtension(path: string): string | null {
   const match = /\.([a-z0-9]+)$/i.exec(path);
