+++ packages/shared/src/review/cacheKeys.ts
@@ -1,0 +1,28 @@
+const FNV_OFFSET_BASIS_32 = 0x811c9dc5;
+const FNV_PRIME_32 = 0x01000193;
+const SECONDARY_HASH_SEED = 0x9e3779b9;
+const SECONDARY_HASH_MULTIPLIER = 0x85ebca6b;
+
+export function fnv1a32(
+  input: string,
+  seed = FNV_OFFSET_BASIS_32,
+  multiplier = FNV_PRIME_32,
+): number {
+  let hash = seed >>> 0;
+  for (let index = 0; index < input.length; index += 1) {
+    hash ^= input.charCodeAt(index);
+    hash = Math.imul(hash, multiplier) >>> 0;
+  }
+  return hash >>> 0;
+}
+
+export function buildPatchCacheKey(patch: string, scope = "diff-panel"): string {
+  const normalizedPatch = patch.trim();
+  const primary = fnv1a32(normalizedPatch, FNV_OFFSET_BASIS_32, FNV_PRIME_32).toString(36);
+  const secondary = fnv1a32(
+    normalizedPatch,
+    SECONDARY_HASH_SEED,
+    SECONDARY_HASH_MULTIPLIER,
+  ).toString(36);
+  return `${scope}:${normalizedPatch.length}:${primary}:${secondary}`;
+}
